+++ src/obfs/verify.h
@@ -14,7 +14,7 @@ struct obfs_t;
 void verify_simple_new_obfs(struct obfs_t * obfs);
 void verify_simple_dispose(struct obfs_t *obfs);
 
-int verify_simple_client_pre_encrypt(struct obfs_t *obfs, char **pplaindata, int datalength, size_t* capacity);
+size_t verify_simple_client_pre_encrypt(struct obfs_t *obfs, char **pplaindata, size_t datalength, size_t* capacity);
 ssize_t verify_simple_client_post_decrypt(struct obfs_t *obfs, char **pplaindata, int datalength, size_t* capacity);
 
 #endif // _OBFS_VERIFY_H
